Remove usage of react-loadable
The `react-loadable` package is generating warnings when building due to
usage of deprecated APIs. There are other alternative packages, but most
of the functionality is built directly into React.

Implement using `React.lazy`, `Suspense`, and `ErrorBoundary` which are
built-in concepts. To avoid a boilerplate `ErrorBoundary`
implementation, the `react-error-boundary` package is used. In the event
that becomes an issue, the implementation is pretty simple.